move FmtModel to a separate file
