+++ src/engine.rs
@@ -1,14 +1,12 @@
 //! This module applies the rules from `super::dsl` to a `SyntaxNode`, to
 //! get a `FmtDiff`.
-use std::collections::HashMap;
+mod fmt_model;
 
-use rnix::{
-    nodes::NODE_ROOT, tokenizer::tokens::TOKEN_WHITESPACE, SmolStr, SyntaxElement, SyntaxNode,
-    SyntaxToken, TextRange, TextUnit,
-};
+use rnix::{tokenizer::tokens::TOKEN_WHITESPACE, SmolStr, SyntaxElement, SyntaxNode, TextRange};
 
 use crate::{
     dsl::{IndentDsl, IndentRule, SpaceLoc, SpaceValue, SpacingDsl, SpacingRule},
+    engine::fmt_model::{BlockPosition, FmtModel, SpaceBlock},
     pattern::{Pattern, PatternSet},
     tree_utils::{has_newline, preceding_tokens, walk_non_whitespace},
     AtomEdit, FmtDiff,
@@ -54,234 +52,6 @@ pub(crate) fn format(
     }
 
     model.into_diff()
-}
-
-/// `FmtModel` is a data structure to which we apply formatting rules.
-///
-/// It wraps a syntax trees and adds `SpaceBlock`s. `SpaceBlock` represents a
-/// run (potentially empty) of whitespace characters. We create whitespace
-/// blocks for existing whitespace tokens. However, if two non-whitespace tokens
-/// are joined together in the syntax tree, we still create an empty
-/// `SpaceBlock` to represent the space between them. That way, rules don't have
-/// to separately handle a case when whitespace node is completely missing from
-/// the original tree.
-///
-/// The `FmtModel` is a mutable data structure, formatting rules work by
-/// changing the actual `SpacingBlock`s. For this reason, the order of
-/// application of the rules is significant.
-///
-/// We maintain the invariant that no two `SpaceBlock`s are directly adjoint to
-/// each other.
-#[derive(Debug)]
-struct FmtModel<'a> {
-    original_node: &'a SyntaxNode,
-    /// We store `SpaceBlock`s in array. With this setup, we can refer to a
-    /// specific block by index, dodging many lifetime issues.
-    blocks: Vec<SpaceBlock<'a>>,
-    /// Maps offset to an index of the block, for which the offset is the start
-    /// offset.
-    by_start_offset: HashMap<TextUnit, usize>,
-    /// Maps offset to an index of the block, for which the offset is the end
-    /// offset.
-    by_end_offset: HashMap<TextUnit, usize>,
-}
-
-#[derive(Debug)]
-struct SpaceBlock<'a> {
-    original: OriginalSpace<'a>,
-    /// Block's textual content, which is seen and modified by formatting rules.
-    new_text: Option<SmolStr>,
-}
-
-#[derive(Debug, Clone, Copy)]
-enum BlockPosition {
-    Before,
-    After,
-}
-
-/// Original whitespace token, if any, that backs a `SpaceBlock.
-#[derive(Debug)]
-enum OriginalSpace<'a> {
-    Some(SyntaxToken<'a>),
-    None { offset: TextUnit },
-}
-
-impl<'a> OriginalSpace<'a> {
-    fn text_range(&self) -> TextRange {
-        match *self {
-            OriginalSpace::Some(token) => token.range(),
-            OriginalSpace::None { offset } => TextRange::from_to(offset, offset),
-        }
-    }
-}
-
-impl<'a> SpaceBlock<'a> {
-    fn new(original: OriginalSpace<'a>) -> SpaceBlock<'a> {
-        SpaceBlock {
-            original,
-            new_text: None,
-        }
-    }
-    fn set_line_break_preserving_existing_newlines(&mut self) {
-        if self.has_newline() {
-            return;
-        }
-        self.set_text("\n");
-    }
-    fn set_text(&mut self, text: &str) {
-        match self.original {
-            OriginalSpace::Some(token) if token.text() == text && self.new_text.is_none() => return,
-            _ => self.new_text = Some(text.into()),
-        }
-    }
-    fn set_indent(&mut self, level: usize) {
-        let indent = " ".repeat(INDENT_SIZE * level);
-        let newlines: String = self.text().chars().filter(|&it| it == '\n').collect();
-        self.set_text(&format!("{}{}", newlines, indent));
-    }
-    fn text(&self) -> &str {
-        if let Some(text) = self.new_text.as_ref() {
-            return text.as_str();
-        }
-        match self.original {
-            OriginalSpace::Some(token) => token.text().as_str(),
-            OriginalSpace::None { .. } => "",
-        }
-    }
-    fn indent_level(&self) -> usize {
-        let text = self.text();
-        match text.rfind('\n') {
-            None => return 0,
-            Some(idx) => text[idx + 1..].chars().count() / INDENT_SIZE,
-        }
-    }
-    fn has_newline(&self) -> bool {
-        self.text().contains('\n')
-    }
-}
-
-impl<'a> FmtModel<'a> {
-    fn new(original_node: &'a SyntaxNode) -> FmtModel<'a> {
-        FmtModel {
-            original_node,
-            blocks: vec![],
-            by_start_offset: HashMap::default(),
-            by_end_offset: HashMap::default(),
-        }
-    }
-
-    fn into_diff(self) -> FmtDiff {
-        let mut diff = FmtDiff {
-            original_node: self.original_node.to_owned(),
-            edits: vec![],
-        };
-        for block in self.blocks {
-            if let Some(new_next) = block.new_text {
-                diff.replace(block.original.text_range(), new_next);
-            }
-        }
-        diff
-    }
-
-    /// This method gets a `SpaceBlock` before or after element. It's pretty
-    /// complicated, because it needs to handle these different cases:
-    /// * We could have already created the block. In this case, we should
-    ///   return the existing block instead of creating a new one.
-    /// * There may, or may not be, backing original whitespace token for the
-    ///   block.
-    /// * The necessary whitespace token is not necessary a sibling of
-    ///   `element`, it might be a sibling of `element`'s ancestor.
-    /// * Finally, root node is special, as it doesn't have siblings and instead
-    ///   leading and trailing whitespace appear as children.
-    fn block_for(
-        &mut self,
-        element: SyntaxElement<'a>,
-        position: BlockPosition,
-    ) -> &mut SpaceBlock<'a> {
-        use BlockPosition::{After, Before};
-
-        assert!(element.kind() != TOKEN_WHITESPACE);
-
-        // Special case, for root node, leading and trailing ws are children of
-        // the root. For all other node types, they are siblings
-        if element.kind() == NODE_ROOT {
-            let root_node = element.as_node().unwrap();
-            let original_space = match position {
-                BlockPosition::Before => root_node.first_child_or_token(),
-                BlockPosition::After => root_node.last_child_or_token(),
-            };
-            return match original_space {
-                Some(SyntaxElement::Token(token)) if token.kind() == TOKEN_WHITESPACE => {
-                    if let Some(&existing) = match position {
-                        Before => self.by_end_offset.get(&token.range().end()),
-                        After => self.by_start_offset.get(&token.range().start()),
-                    } {
-                        &mut self.blocks[existing]
-                    } else {
-                        self.push_block(SpaceBlock::new(OriginalSpace::Some(token)))
-                    }
-                }
-                _ => {
-                    let offset = match position {
-                        Before => root_node.range().start(),
-                        After => root_node.range().end(),
-                    };
-
-                    if let Some(&existing) = match position {
-                        Before => self.by_end_offset.get(&offset),
-                        After => self.by_start_offset.get(&offset),
-                    } {
-                        &mut self.blocks[existing]
-                    } else {
-                        self.push_block(SpaceBlock::new(OriginalSpace::None { offset }))
-                    }
-                }
-            };
-        }
-
-        let offset = match position {
-            Before => element.range().start(),
-            After => element.range().end(),
-        };
-
-        if let Some(&existing) = match position {
-            Before => self.by_end_offset.get(&offset),
-            After => self.by_start_offset.get(&offset),
-        } {
-            return &mut self.blocks[existing];
-        }
-
-        let original_token = match position {
-            Before => element.prev_sibling_or_token(),
-            After => element.next_sibling_or_token(),
-        };
-
-        let original_space = match original_token {
-            Some(SyntaxElement::Token(token)) if token.kind() == TOKEN_WHITESPACE => {
-                OriginalSpace::Some(token)
-            }
-            Some(_) => OriginalSpace::None { offset },
-            _ => match element.parent() {
-                Option::Some(parent) => return self.block_for(parent.into(), position),
-                None => OriginalSpace::None { offset },
-            },
-        };
-
-        self.push_block(SpaceBlock::new(original_space))
-    }
-
-    fn push_block(&mut self, block: SpaceBlock<'a>) -> &mut SpaceBlock<'a> {
-        let idx = self.blocks.len();
-        let range = block.original.text_range();
-
-        let prev = self.by_start_offset.insert(range.start(), idx);
-        assert!(prev.is_none());
-        let prev = self.by_end_offset.insert(range.end(), idx);
-        assert!(prev.is_none());
-
-        self.blocks.push(block);
-        self.blocks.last_mut().unwrap()
-    }
 }
 
 impl SpacingRule {
